chaging distance for integers

+++ ConstantDistanceSet.py
@@ -40,12 +40,12 @@ class ConstantDistanceSet(Set):
 
                     # Extra condition: we need to be able to split the distance in a logical way for the segments
                     if distance/nBlocks/globals.minSegmentDistance == np.floor(distance/nBlocks/globals.minSegmentDistance):
-                        optionIncreaseDecreaseSplit[distance] = np.arange(start=distance/nBlocks, stop=(nBlocks+1)*distance/nBlocks, step=distance/nBlocks)
+                        optionIncreaseDecreaseSplit[distance] = np.arange(start=distance/nBlocks, stop=int((nBlocks+1)*distance/nBlocks), step=int(distance/nBlocks))
 
                     # We also accept cases where we add a "0" at the start of the sequence
                     if nBlocks != 1:
                         if distance/(nBlocks-1)/globals.minSegmentDistance == np.floor(distance/(nBlocks-1)/globals.minSegmentDistance):
-                            optionIncreaseDecreaseSplit[distance] = np.arange(start=0, stop=(nBlocks)*distance/(nBlocks-1), step=distance/(nBlocks-1)) 
+                            optionIncreaseDecreaseSplit[distance] = np.arange(start=0, stop=int((nBlocks)*distance/(nBlocks-1)), step=int(distance/(nBlocks-1))) 
 
         # Then deciding what type of Block sequence this will be: 
             # randomSplit: The block will have a random number of segments that will simply repeat from one block to the other
